test: increased test coverage
1.5/1.6 overrides are now conditionally loaded depending on the actual
Elixir version
